+++ application/tests/_ci_phpunit_test/CIPHPUnitTestRouter.php
@@ -55,6 +55,14 @@ class CIPHPUnitTestRouter
 
 		if ($e404)
 		{
+			// If 404, CodeIgniter instance is not created yet. So create it here.
+			// Because we need CI->output->_status
+			$CI =& get_instance();
+			if ($CI instanceof CIPHPUnitTestNullCodeIgniter)
+			{
+				new CI_Controller();
+			}
+
 			show_404($RTR->directory.$class.'/'.$method.' is not found');
 		}